Add closeness tracking with covariance scoring and Relationships tab
Track member affinity from per-proposal support patterns using a
covariance-based score (agreement on niche proposals matters; agreement
on unanimous proposals does not). Computed once per proposal at pulse
execution and stored as float in closeness_records.

Viewer gains a Relationships tab with three views: a sorted adjacency
heatmap (default, scales to ~80 members), a classical-MDS scatter
("Map", scales to hundreds), and the original force-directed graph for
small communities.

+++ src/kbz/models/closeness.py
@@ -1,7 +1,7 @@
 import uuid
 from datetime import datetime
 
-from sqlalchemy import Integer, DateTime, text
+from sqlalchemy import Float, DateTime, text
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import Mapped, mapped_column
 
@@ -13,5 +13,5 @@ class Closeness(Base):
 
     user_id1: Mapped[uuid.UUID] = mapped_column(UUID(as_uuid=True), primary_key=True)
     user_id2: Mapped[uuid.UUID] = mapped_column(UUID(as_uuid=True), primary_key=True)
-    score: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
+    score: Mapped[float] = mapped_column(Float, nullable=False, default=0.0)
     last_calculation: Mapped[datetime] = mapped_column(DateTime(timezone=True), nullable=False, server_default=text("NOW()"))
